assignment4/requirements.txt
notebook==6.1.3
numpy==1.19.1
matplotlib==3.3.1
gym==0.17.3
pymdptoolbox==4.0-b3
